Implement file IO in Inko
This replaces the use of Rust code for file IO with Inko code using the
C FFI. This means that for file IO we no longer depend on Rust semantics
and its lack of a stable ABI, and have more control over how we want to
implement the standard library. Not having to heap allocate various
objects (due to the lack of an ABI) should also improve performance a
little bit.

This commit includes changes related to defining the types and values of
constants. Prior to this commit the work was performed in two passes,
but this approach wasn't able to handle certain cases introduced by the
IO changes. In the new setup, both defining the types and values uses a
work list that is processed until no more constants remain, or until we
detect a circular dependency (e.g. A = B and B = A).

This fixes https://github.com/inko-lang/inko/issues/633 and closes
https://github.com/inko-lang/inko/issues/583.

Changelog: changed

+++ rt/src/runtime/general.rs
@@ -53,8 +53,8 @@ pub unsafe extern "system" fn inko_alloc_error(class: ClassPointer) -> ! {
 }
 
 #[no_mangle]
-pub unsafe extern "system" fn inko_last_error() -> i32 {
-    Error::last_os_error().raw_os_error().unwrap_or(-1)
+pub unsafe extern "system" fn inko_last_error() -> i64 {
+    Error::last_os_error().raw_os_error().unwrap_or(-1) as i64
 }
 
 #[no_mangle]
